requirements.txt
# AI provider SDKs
openai>=1.0.0,<2.0.0
anthropic>=0.25.0,<1.0.0
google-generativeai>=0.5.0,<1.0.0

# Dashboard
streamlit>=1.30.0,<2.0.0
pandas>=2.0.0,<3.0.0

# Database
psycopg2-binary>=2.9.0,<3.0.0

# Resilience
tenacity>=8.0.0,<10.0.0
